pyproject.toml
[project]
name = "rationai-mlkit"
version = "0.1.14"
description = ""
authors = [
    { name = "Matěj Pekár", email = "525077@mail.muni.cz" },
    { name = "Jakub Pekár", email = "jakubpekar@mail.muni.cz" },
]
requires-python = ">=3.11,<3.12"
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
    "lightning>=2.2.4",
    "GitPython>=3.1.43",
    "mlflow>=2.12.1",
    "ray[train]>=2.20.0",
]

[tool.pdm.dev-dependencies]
dev = ["mypy", "pre-commit", "ruff"]
test = ["pytest", "pytest-cov"]


[tool.pdm.scripts]
l = { composite = ["lint", "format", "mypy"] }
test = "pytest tests"
lint = "ruff check"
format = "ruff format"
mypy = "mypy ."
post_install = { composite = [
    "pre-commit install",
    "pre-commit install --hook-type commit-msg",
] }
